Rename files in test code for better search
