+++ rust/main/lander/src/adapter/chains/sealevel/adapter/tests.rs
@@ -1,7 +1,7 @@
-mod build;
-mod common;
-mod config;
-mod estimate;
-mod simulate;
-mod status;
-mod submit;
+mod tests_build;
+mod tests_common;
+mod tests_config;
+mod tests_estimate;
+mod tests_simulate;
+mod tests_status;
+mod tests_submit;
