+++ rust/main/lander/src/adapter/chains/sealevel/adapter/tests/tests_submit.rs
@@ -1,6 +1,6 @@
 use crate::adapter::{
     chains::sealevel::{
-        adapter::tests::common::{adapter, payload, precursor},
+        adapter::tests::tests_common::{adapter, payload, precursor},
         transaction::TransactionFactory,
     },
     AdaptsChain,
